+++ aicir/qas/algorithms/MoG_VQE.py
@@ -42,7 +42,7 @@ import numpy as np
 
 from ...backends.numpy_backend import NumpyBackend
 from ...core.circuit import Circuit, cx, ry, rz
-from ...operators import Hamiltonian
+from ...core.operators import Hamiltonian
 
 
 Edge = tuple[int, int]